fix types

+++ src/main.c
@@ -52,7 +52,7 @@ int Menu_Main(void)
 
     {
         VPADData vpad;
-        long int vpadError = -1;
+        s32 vpadError = -1;
         VPADRead(0, &vpad, 1, &vpadError);
 
         if(vpadError == 0)
